requirements.txt
aiofiles==23.2.1
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==3.7.1
apt-clone==0.2.1
apturl==0.5.2
async-timeout==4.0.1
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
email_validator==2.2.0
exceptiongroup==1.3.0
eyeD3==0.8.10
fastapi==0.104.1
ffmpeg-python==0.2.0
filelock==3.6.0
future==1.0.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.25.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mutagen==1.47.0
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
onboard==1.4.1
opencv-python-headless==4.8.1.78
orjson==3.9.10
packaging==21.3
PAM==0.4.2
pathlib2==2.3.7
pexpect==4.8.0
Pillow==10.1.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyelftools==0.27
pygame==2.5.2
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-gnupg==0.4.8
python-magic==0.4.24
python-multipart==0.0.6
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
ssh-import-id==5.11
starlette==0.27.0
systemd-python==234
fastapi
uvicorn
psutil
jinja2
tinycss2==1.1.1
tldextract==3.1.2
typing-inspection==0.4.1
typing_extensions==4.14.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
Unidecode==1.3.3
urllib3==1.26.5
uvicorn==0.24.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
